#80 moved obj utilities to core
