+++ allensdk/api/cache.py
@@ -38,7 +38,6 @@ from allensdk.config.manifest_builder import ManifestBuilder
 import allensdk.core.json_utilities as ju
 from allensdk.deprecated import deprecated
 
-import numpy as np
 import pandas as pd
 import pandas.io.json as pj
 
@@ -598,68 +597,3 @@ def cacheable(strategy=None,
         
         return w
     return decor
-
-
-def read_obj(path):
-    with open(path, 'r') as obj_file:
-        lines = obj_file.read().split('\n')
-        output = parse_obj(lines)
-    return output
-
-
-def parse_obj(lines):
-    '''Parse a wavefront obj file into a triplet of vertices, normals, and faces. 
-    This parser is specific to obj files generated from our annotation volumes
-
-    Parameters
-    ----------
-    lines : list of str
-        Lines of input obj file
-
-    Returns
-    -------
-    vertices : np.ndarray
-        Dimensions are (nSamples, nCoordinates=3). Locations in the reference space
-        of vertices
-    vertex_normals : np.ndarray
-        Dimensions are (nSample, nElements=3). Vectors normal to vertices.
-    face_vertices : np.ndarray
-        Dimensions are (sample, nVertices=3). References are given in indices 
-        (0-indexed here, but 1-indexed in the file) of vertices that make up each face.
-    face_normals : np.ndarray
-        Dimensions are (sample, nNormals=3). References are given in indices 
-        (0-indexed here, but 1-indexed in the file) of vertex normals that make up each face.
-
-    Notes
-    -----
-    This parser is specialized to the obj files that the Allen Institute for Brain Science 
-    generates from our own structure annotations.
-    '''
-
-    vertices = []
-    vertex_normals = []
-    face_vertices = []
-    face_normals = []
-
-    for line in lines:
-        
-        print line[:2] == 'f '
-
-        if line[:2] == 'v ':
-            vertices.append( line.split()[1:] )
-
-        elif line[:3] == 'vn ':
-            vertex_normals.append( line.split()[1:] )
-
-        elif line[:2] == 'f ':
-            line = line.replace('//', ' ').split()[1:]
-
-            face_vertices.append( line[::2] )
-            face_normals.append( line[1::2] )
-            
-    vertices = np.array(vertices).astype(float)
-    vertex_normals = np.array(vertex_normals).astype(float)
-    face_vertices = np.array(face_vertices).astype(int) - 1
-    face_normals = np.array(face_normals).astype(int) - 1
-
-    return vertices, vertex_normals, face_vertices, face_normals
